# Swap to a new faster rolling hash algorithm.