--- a/src/chunker.rs
+++ b/src/chunker.rs
@@ -1,8 +1,7 @@
-use super::rollsum::{GearTab, Rollsum, WINDOW_SIZE};
+use super::rollsum::{FastGearHasher, GearTab, RollsumSplitter};
 
 pub struct RollsumChunker {
-    gear_tab: GearTab,
-    rs: Rollsum,
+    rs: FastGearHasher,
     min_sz: usize,
     max_sz: usize,
     default_chunk_capacity: usize,
@@ -19,8 +18,7 @@ impl RollsumChunker {
         }
         let default_chunk_capacity = max_sz / 2;
         RollsumChunker {
-            rs: Rollsum::new(),
-            gear_tab,
+            rs: FastGearHasher::new(gear_tab),
             min_sz,
             max_sz,
             default_chunk_capacity,
@@ -39,44 +37,42 @@ impl RollsumChunker {
     }
 
     pub fn add_bytes(&mut self, buf: &[u8]) -> (usize, Option<Vec<u8>>) {
-        debug_assert!(self.cur_vec.len() < self.max_sz);
-
         let mut n_bytes = buf.len();
 
         if (n_bytes + self.cur_vec.len()) > self.max_sz {
             let overshoot = (n_bytes + self.cur_vec.len()) - self.max_sz;
             n_bytes -= overshoot;
-            debug_assert!(self.cur_vec.len() + n_bytes <= self.max_sz);
         }
 
         if self.spare_capacity() < n_bytes {
-            let mut growth = self.max_sz / 3;
-            if growth == 0 {
-                growth = 1;
-            }
+            let mut growth = (self.max_sz / 3).max(1);
             if self.cur_vec.capacity() + growth > self.max_sz {
                 growth = self.max_sz - self.cur_vec.capacity();
             }
             self.cur_vec.reserve(growth);
-            debug_assert!(self.spare_capacity() >= n_bytes);
+            n_bytes = std::cmp::min(self.spare_capacity(), n_bytes);
         }
 
         // None of the bytes we are adding will count towards the
         // next chunk, simply add them all, the bytes don't matter
         // as we will cycle WINDOW_SIZE too.
-        if self.min_sz >= WINDOW_SIZE
-            && (self.cur_vec.len() + n_bytes < (self.min_sz - WINDOW_SIZE))
-        {
-            self.cur_vec.extend_from_slice(&buf[0..n_bytes]);
-            return (n_bytes, None);
+        if let Some(window_size) = self.rs.window_size() {
+            if self.min_sz >= window_size
+                && (self.cur_vec.len() + n_bytes < (self.min_sz - window_size))
+            {
+                self.cur_vec.extend_from_slice(&buf[0..n_bytes]);
+                return (n_bytes, None);
+            }
         }
 
-        n_bytes = std::cmp::min(self.spare_capacity(), n_bytes);
-
-        match self.rs.roll_bytes(&self.gear_tab, &buf[0..n_bytes]) {
+        match self.rs.roll_bytes(&buf[0..n_bytes]) {
             Some(split) => {
                 self.cur_vec.extend_from_slice(&buf[0..split]);
-                (split, Some(self.swap_vec()))
+                if self.cur_vec.len() < self.min_sz {
+                    (split, None)
+                } else {
+                    (split, Some(self.swap_vec()))
+                }
             }
             None => {
                 self.cur_vec.extend_from_slice(&buf[0..n_bytes]);
@@ -116,12 +112,12 @@ impl RollsumChunker {
 
 #[cfg(test)]
 mod tests {
-    use super::super::rollsum::TEST_GEAR_TAB;
+    use super::super::rollsum::{GearTab, TEST_GEAR_TAB_DATA};
     use super::*;
 
     #[test]
     fn test_add_bytes() {
-        let mut ch = RollsumChunker::new(TEST_GEAR_TAB, 1, 2);
+        let mut ch = RollsumChunker::new(GearTab::from_array(TEST_GEAR_TAB_DATA), 1, 2);
 
         match ch.add_bytes(b"a") {
             (1, None) => (),
@@ -143,7 +139,7 @@ mod tests {
 
     #[test]
     fn test_force_split_bytes() {
-        let mut ch = RollsumChunker::new(TEST_GEAR_TAB, 10, 100);
+        let mut ch = RollsumChunker::new(GearTab::from_array(TEST_GEAR_TAB_DATA), 10, 100);
         assert_eq!(ch.force_split(), None);
         ch.add_bytes(b"abc");