This implements an entirely new rolling hash algorithm which we have called 'interleaved gear hash'. As far as we are aware the algorithm was invented by Quentin Carbonneaux and is currently unpublished. The change also includes SIMD implementations of these hash functions, although they currently require nightly rust to use.

Benchmarks:
```
Time to hash 5 GiB

export PATH="$PWD/target/release:$PATH"

export data="dd if=/dev/zero bs=1M obs=1M count=$((5*1024))"

hyperfine \
  --warmup 1 \
  "\$data | bupstash rollsum-benchmark 'GearHasher'" \
  "\$data | bupstash rollsum-benchmark 'InterleavedGearHasher<2>'" \
  "\$data | bupstash rollsum-benchmark 'InterleavedGearHasher<4>'" \
  "\$data | bupstash rollsum-benchmark 'InterleavedGearHasher<8>'" \
  "\$data | bupstash rollsum-benchmark 'InterleavedGearHasher<16>'" \
  "\$data | bupstash rollsum-benchmark 'SimdInterleavedGearHasher<2>'" \
  "\$data | bupstash rollsum-benchmark 'SimdInterleavedGearHasher<4>'" \
  "\$data | bupstash rollsum-benchmark 'SimdInterleavedGearHasher<8>'" \
  "\$data | bupstash rollsum-benchmark 'SimdInterleavedGearHasher<16>'"

# M1 Mac

Benchmark 1: $data | bupstash rollsum-benchmark 'GearHasher'
  Time (mean ± σ):      3.986 s ±  0.013 s    [User: 3.569 s, System: 0.758 s]
  Range (min … max):    3.961 s …  4.007 s    10 runs

Benchmark 2: $data | bupstash rollsum-benchmark 'InterleavedGearHasher<2>'
  Time (mean ± σ):      2.990 s ±  0.039 s    [User: 2.289 s, System: 0.823 s]
  Range (min … max):    2.938 s …  3.064 s    10 runs

Benchmark 3: $data | bupstash rollsum-benchmark 'InterleavedGearHasher<4>'
  Time (mean ± σ):      2.481 s ±  0.048 s    [User: 1.798 s, System: 0.786 s]
  Range (min … max):    2.382 s …  2.568 s    10 runs

Benchmark 4: $data | bupstash rollsum-benchmark 'InterleavedGearHasher<8>'
  Time (mean ± σ):      2.461 s ±  0.021 s    [User: 1.767 s, System: 0.799 s]
  Range (min … max):    2.426 s …  2.505 s    10 runs

Benchmark 5: $data | bupstash rollsum-benchmark 'InterleavedGearHasher<16>'
  Time (mean ± σ):      3.138 s ±  0.142 s    [User: 2.283 s, System: 0.882 s]
  Range (min … max):    2.939 s …  3.361 s    10 runs

Benchmark 6: $data | bupstash rollsum-benchmark 'SimdInterleavedGearHasher<2>'
  Time (mean ± σ):      4.198 s ±  0.036 s    [User: 3.751 s, System: 0.821 s]
  Range (min … max):    4.158 s …  4.270 s    10 runs

Benchmark 7: $data | bupstash rollsum-benchmark 'SimdInterleavedGearHasher<4>'
  Time (mean ± σ):      2.570 s ±  0.053 s    [User: 1.893 s, System: 0.801 s]
  Range (min … max):    2.506 s …  2.667 s    10 runs

Benchmark 8: $data | bupstash rollsum-benchmark 'SimdInterleavedGearHasher<8>'
  Time (mean ± σ):      2.075 s ±  0.107 s    [User: 1.409 s, System: 0.791 s]
  Range (min … max):    1.872 s …  2.239 s    10 runs

Benchmark 9: $data | bupstash rollsum-benchmark 'SimdInterleavedGearHasher<16>'
  Time (mean ± σ):      2.066 s ±  0.110 s    [User: 1.392 s, System: 0.793 s]
  Range (min … max):    1.837 s …  2.243 s    10 runs

Summary
  '$data | bupstash rollsum-benchmark 'SimdInterleavedGearHasher<16>'' ran
    1.00 ± 0.07 times faster than '$data | bupstash rollsum-benchmark 'SimdInterleavedGearHasher<8>''
    1.19 ± 0.06 times faster than '$data | bupstash rollsum-benchmark 'InterleavedGearHasher<8>''
    1.20 ± 0.07 times faster than '$data | bupstash rollsum-benchmark 'InterleavedGearHasher<4>''
    1.24 ± 0.07 times faster than '$data | bupstash rollsum-benchmark 'SimdInterleavedGearHasher<4>''
    1.45 ± 0.08 times faster than '$data | bupstash rollsum-benchmark 'InterleavedGearHasher<2>''
    1.52 ± 0.11 times faster than '$data | bupstash rollsum-benchmark 'InterleavedGearHasher<16>''
    1.93 ± 0.10 times faster than '$data | bupstash rollsum-benchmark 'GearHasher''
    2.03 ± 0.11 times faster than '$data | bupstash rollsum-benchmark 'SimdInterleavedGearHasher<2>''

# AMD Ryzen

Benchmark 1: $data | bupstash rollsum-benchmark 'GearHasher'
  Time (mean ± σ):      3.735 s ±  0.042 s    [User: 2.869 s, System: 2.161 s]
  Range (min … max):    3.682 s …  3.802 s    10 runs

Benchmark 2: $data | bupstash rollsum-benchmark 'InterleavedGearHasher<2>'
  Time (mean ± σ):      4.197 s ±  0.050 s    [User: 3.392 s, System: 2.177 s]
  Range (min … max):    4.103 s …  4.253 s    10 runs

Benchmark 3: $data | bupstash rollsum-benchmark 'InterleavedGearHasher<4>'
  Time (mean ± σ):      3.225 s ±  0.067 s    [User: 2.297 s, System: 2.142 s]
  Range (min … max):    3.052 s …  3.287 s    10 runs

Benchmark 4: $data | bupstash rollsum-benchmark 'InterleavedGearHasher<8>'
  Time (mean ± σ):      3.718 s ±  1.900 s    [User: 2.812 s, System: 2.143 s]
  Range (min … max):    3.075 s …  9.125 s    10 runs

  Warning: Statistical outliers were detected. Consider re-running this benchmark on a quiet PC without any interferences from other programs. It might help to use the '--warmup' or '--prepare' options.

Benchmark 5: $data | bupstash rollsum-benchmark 'InterleavedGearHasher<16>'
  Time (mean ± σ):      4.044 s ±  0.044 s    [User: 3.235 s, System: 2.154 s]
  Range (min … max):    4.005 s …  4.154 s    10 runs

Benchmark 6: $data | bupstash rollsum-benchmark 'SimdInterleavedGearHasher<2>'
  Time (mean ± σ):      3.194 s ±  0.064 s    [User: 2.263 s, System: 2.152 s]
  Range (min … max):    3.061 s …  3.270 s    10 runs

Benchmark 7: $data | bupstash rollsum-benchmark 'SimdInterleavedGearHasher<4>'
  Time (mean ± σ):      2.711 s ±  0.009 s    [User: 1.684 s, System: 2.189 s]
  Range (min … max):    2.696 s …  2.725 s    10 runs

Benchmark 8: $data | bupstash rollsum-benchmark 'SimdInterleavedGearHasher<8>'
  Time (mean ± σ):      2.722 s ±  0.063 s    [User: 1.714 s, System: 2.166 s]
  Range (min … max):    2.567 s …  2.799 s    10 runs

Benchmark 9: $data | bupstash rollsum-benchmark 'SimdInterleavedGearHasher<16>'
  Time (mean ± σ):      2.737 s ±  0.029 s    [User: 1.722 s, System: 2.179 s]
  Range (min … max):    2.702 s …  2.783 s    10 runs

Summary
  '$data | bupstash rollsum-benchmark 'SimdInterleavedGearHasher<4>'' ran
    1.00 ± 0.02 times faster than '$data | bupstash rollsum-benchmark 'SimdInterleavedGearHasher<8>''
    1.01 ± 0.01 times faster than '$data | bupstash rollsum-benchmark 'SimdInterleavedGearHasher<16>''
    1.18 ± 0.02 times faster than '$data | bupstash rollsum-benchmark 'SimdInterleavedGearHasher<2>''
    1.19 ± 0.03 times faster than '$data | bupstash rollsum-benchmark 'InterleavedGearHasher<4>''
    1.37 ± 0.70 times faster than '$data | bupstash rollsum-benchmark 'InterleavedGearHasher<8>''
    1.38 ± 0.02 times faster than '$data | bupstash rollsum-benchmark 'GearHasher''
    1.49 ± 0.02 times faster than '$data | bupstash rollsum-benchmark 'InterleavedGearHasher<16>''
    1.55 ± 0.02 times faster than '$data | bupstash rollsum-benchmark 'InterleavedGearHasher<2>''
```